redefined folder structures (creator)
