--- pocketHero/pocketHero/src/main/java/com/example/pocketHero/domains/Player.java
@@ -1,5 +1,0 @@
-package com.example.pocketHero.domains;
-
-public class Player {
-    
-}
