+++ pocketHero/pocketHero/src/main/java/com/example/pocketHero/domains/creator/Campaign.java
@@ -1,4 +1,4 @@
-package com.example.pocketHero.domains;
+package com.example.pocketHero.domains.creator;
 
 import java.util.List;
