+++ pocketHero/pocketHero/src/main/java/com/example/pocketHero/domains/creator/Player.java
@@ -1,0 +1,5 @@
+package com.example.pocketHero.domains.creator;
+
+public class Player {
+    
+}
