+++ pocketHero/pocketHero/src/main/java/com/example/pocketHero/domains/creator/Stadistics.java
@@ -1,4 +1,4 @@
-package com.example.pocketHero.domains;
+package com.example.pocketHero.domains.creator;
 
 public class Stadistics {
